requirements.txt
flask
flask-cors
groq
supabase
python-dotenv
